[UPDATE]: remove "robot" from the intro
Although this is related somewhat to robotics, I get the point that other colleagues might not consider it robotics, because when dealing
with actual robots, we end up working with a bit harder problems (have to deal with real-world situations). We try to study how to solve these issues with simulation, so I hope this work can give some light in the matter.